+++ src/actions/users.js
@@ -3,35 +3,6 @@ import {SubmissionError} from 'redux-form';
 import {API_BASE_URL} from '../config';
 import {normalizeResponseErrors} from './utils';
 
-export const fetchUserFriendlist = user => (dispatch, getState) => {
-    const authToken = getState().auth.authToken;
-    return fetch(`${API_BASE_URL}/users/${user}`, {
-        method: 'GET',
-        headers: {
-            'content-type': 'application/json',
-            Authorization: `Bearer ${authToken}`
-        }
-    })
-    .then(res => normalizeResponseErrors(res))
-    .then(res => res.json())
-    .catch(err => console.log(err))
-}
-
-//SUGGESTED FRIENDS LIST?
-export const fetchUserSuggestedlist = user => (dispatch, getState) => {
-    const authToken = getState().auth.authToken;
-    return fetch(`${API_BASE_URL}/users/${user}`, {
-        method: 'GET',
-        headers: {
-            'content-type': 'application/json',
-            Authorization: `Bearer ${authToken}`
-        }
-    })
-    .then(res => normalizeResponseErrors(res))
-    .then(res => res.json())
-    .catch(err => console.log(err))
-}
-
 
 export const registerUser = user => dispatch => {
     return fetch(`${API_BASE_URL}/users`, {
@@ -56,6 +27,3 @@ export const registerUser = user => dispatch => {
         });
 };
 
-export const fetchUserFriendlist = () = (dispatch, getState) => {
-
-};
