+++ src/reducers/friends.js
@@ -1,0 +1,5 @@
+import {
+  FETCH_USER_FRIENDS_REQUEST,
+  FETCH_USER_FRIENDS_SUCCESS,
+  FETCH_USER_FRIENDS_ERROR
+}from '../actions/users';
